Split equal algorithm unit tests into header and source
This enables to use existing facilites for datapar mismatch unit tests
