+++ libs/core/algorithms/tests/unit/algorithms/equal_tests.hpp
@@ -1,0 +1,455 @@
+//  Copyright (c) 2014-2020 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#pragma once
+
+#include <hpx/modules/testing.hpp>
+#include <hpx/parallel/algorithms/equal.hpp>
+
+#include <cstddef>
+#include <iostream>
+#include <iterator>
+#include <numeric>
+#include <random>
+#include <string>
+#include <vector>
+
+#include "test_utils.hpp"
+
+///////////////////////////////////////////////////////////////////////////////
+unsigned int seed = std::random_device{}();
+std::mt19937 gen(seed);
+
+template <typename IteratorTag>
+void test_equal1(IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    {
+        bool result = hpx::equal(
+            iterator(std::begin(c1)), iterator(std::end(c1)), std::begin(c2));
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+
+    {
+        std::uniform_int_distribution<> dis(0, c1.size() - 1);
+        ++c1[dis(gen)];    //-V104
+        bool result = hpx::equal(
+            iterator(std::begin(c1)), iterator(std::end(c1)), std::begin(c2));
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_equal1(ExPolicy&& policy, IteratorTag)
+{
+    static_assert(hpx::is_execution_policy<ExPolicy>::value,
+        "hpx::is_execution_policy<ExPolicy>::value");
+
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    {
+        bool result = hpx::equal(policy, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2));
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+
+    {
+        std::uniform_int_distribution<> dis(0, c1.size() - 1);
+        ++c1[dis(gen)];    //-V104
+        bool result = hpx::equal(policy, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2));
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_equal1_async(ExPolicy&& p, IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    {
+        hpx::future<bool> result = hpx::equal(p, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2));
+        result.wait();
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result.get(), expected);
+    }
+
+    {
+        std::uniform_int_distribution<> dis(0, c1.size() - 1);
+        ++c1[dis(gen)];    //-V104
+
+        hpx::future<bool> result = hpx::equal(p, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2));
+        result.wait();
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result.get(), expected);
+    }
+}
+
+///////////////////////////////////////////////////////////////////////////////
+template <typename IteratorTag>
+void test_equal2(IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    {
+        bool result = hpx::equal(iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::equal_to<>());
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+
+    {
+        std::uniform_int_distribution<> dis(0, c1.size() - 1);
+        ++c1[dis(gen)];    //-V104
+        bool result = hpx::equal(iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::equal_to<>());
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_equal2(ExPolicy&& policy, IteratorTag)
+{
+    static_assert(hpx::is_execution_policy<ExPolicy>::value,
+        "hpx::is_execution_policy<ExPolicy>::value");
+
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    {
+        bool result = hpx::equal(policy, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::equal_to<>());
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+
+    {
+        std::uniform_int_distribution<> dis(0, c1.size() - 1);
+        ++c1[dis(gen)];    //-V104
+        bool result = hpx::equal(policy, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::equal_to<>());
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result, expected);
+    }
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_equal2_async(ExPolicy&& p, IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    {
+        hpx::future<bool> result = hpx::equal(p, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::equal_to<>());
+        result.wait();
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result.get(), expected);
+    }
+
+    {
+        std::uniform_int_distribution<> dis(0, c1.size() - 1);
+        ++c1[dis(gen)];    //-V104
+
+        hpx::future<bool> result = hpx::equal(p, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::equal_to<>());
+        result.wait();
+
+        bool expected =
+            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+
+        // verify values
+        HPX_TEST_EQ(result.get(), expected);
+    }
+}
+
+///////////////////////////////////////////////////////////////////////////////
+template <typename IteratorTag>
+void test_equal_exception(IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    bool caught_exception = false;
+    try
+    {
+        hpx::equal(iterator(std::begin(c1)), iterator(std::end(c1)),
+            std::begin(c2),
+            [](auto, auto) { return throw std::runtime_error("test"), true; });
+
+        HPX_TEST(false);
+    }
+    catch (hpx::exception_list const& e)
+    {
+        caught_exception = true;
+        test::test_num_exceptions<hpx::execution::sequenced_policy,
+            IteratorTag>::call(hpx::execution::seq, e);
+    }
+    catch (...)
+    {
+        HPX_TEST(false);
+    }
+
+    HPX_TEST(caught_exception);
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_equal_exception(ExPolicy&& policy, IteratorTag)
+{
+    static_assert(hpx::is_execution_policy<ExPolicy>::value,
+        "hpx::is_execution_policy<ExPolicy>::value");
+
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    bool caught_exception = false;
+    try
+    {
+        hpx::equal(policy, iterator(std::begin(c1)), iterator(std::end(c1)),
+            std::begin(c2),
+            [](auto, auto) { return throw std::runtime_error("test"), true; });
+
+        HPX_TEST(false);
+    }
+    catch (hpx::exception_list const& e)
+    {
+        caught_exception = true;
+        test::test_num_exceptions<ExPolicy, IteratorTag>::call(policy, e);
+    }
+    catch (...)
+    {
+        HPX_TEST(false);
+    }
+
+    HPX_TEST(caught_exception);
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_equal_exception_async(ExPolicy&& p, IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    bool caught_exception = false;
+    bool returned_from_algorithm = false;
+    try
+    {
+        hpx::future<bool> f = hpx::equal(p, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2),
+            [](auto, auto) { return throw std::runtime_error("test"), true; });
+        returned_from_algorithm = true;
+        f.get();
+
+        HPX_TEST(false);
+    }
+    catch (hpx::exception_list const& e)
+    {
+        caught_exception = true;
+        test::test_num_exceptions<ExPolicy, IteratorTag>::call(p, e);
+    }
+    catch (...)
+    {
+        HPX_TEST(false);
+    }
+
+    HPX_TEST(caught_exception);
+    HPX_TEST(returned_from_algorithm);
+}
+
+///////////////////////////////////////////////////////////////////////////////
+template <typename ExPolicy, typename IteratorTag>
+void test_equal_bad_alloc(ExPolicy&& policy, IteratorTag)
+{
+    static_assert(hpx::is_execution_policy<ExPolicy>::value,
+        "hpx::is_execution_policy<ExPolicy>::value");
+
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    bool caught_bad_alloc = false;
+    try
+    {
+        hpx::equal(policy, iterator(std::begin(c1)), iterator(std::end(c1)),
+            std::begin(c2),
+            [](auto, auto) { return throw std::bad_alloc(), true; });
+
+        HPX_TEST(false);
+    }
+    catch (std::bad_alloc const&)
+    {
+        caught_bad_alloc = true;
+    }
+    catch (...)
+    {
+        HPX_TEST(false);
+    }
+
+    HPX_TEST(caught_bad_alloc);
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_equal_bad_alloc_async(ExPolicy&& p, IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c1(10007);
+    std::vector<std::size_t> c2(c1.size());
+
+    std::size_t first_value = gen();    //-V101
+    std::iota(std::begin(c1), std::end(c1), first_value);
+    std::iota(std::begin(c2), std::end(c2), first_value);
+
+    bool caught_bad_alloc = false;
+    bool returned_from_algorithm = false;
+    try
+    {
+        hpx::future<bool> f = hpx::equal(p, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2),
+            [](auto, auto) { return throw std::bad_alloc(), true; });
+        returned_from_algorithm = true;
+        f.get();
+
+        HPX_TEST(false);
+    }
+    catch (std::bad_alloc const&)
+    {
+        caught_bad_alloc = true;
+    }
+    catch (...)
+    {
+        HPX_TEST(false);
+    }
+
+    HPX_TEST(caught_bad_alloc);
+    HPX_TEST(returned_from_algorithm);
+}
